Library

Documentation for SatelliteToolboxPropagators.jl.

SatelliteToolboxPropagators.J2OsculatingPropagator — Type

mutable struct J2OsculatingPropagator{Tepoch<:Number, T<:Number}

J2 osculating orbit propagator structure.

SatelliteToolboxPropagators.J2Propagator — Type

mutable struct J2Propagator{Tepoch<:Number, T<:Number}

J2 orbit propagator structure.

SatelliteToolboxPropagators.J2PropagatorConstants — Type

struct J2PropagatorConstants{T<:Number}

Constants for the J2 orbit propagator.

Fields

R0::T: Earth equatorial radius [m].

μm::T: √(GM / R0^3) [er/s]^(3/2).

J2::T: The second gravitational zonal harmonic of the Earth.

SatelliteToolboxPropagators.J4Propagator — Type

J4Propagator{Tepoch, T}

J4 orbit propagator structure.

SatelliteToolboxPropagators.J4PropagatorConstants — Type

struct J4PropagatorConstants{T<:Number}

Constants for the J4 orbit propagator.

Fields

R0::T: Earth equatorial radius [m].

μm::T: √(GM / R0^3) [er/s]^(3/2).

J2::T: The second gravitational zonal harmonic of the Earth.

J4::T: The fourth gravitational zonal harmonic of the Earth.

SatelliteToolboxPropagators.OrbitPropagatorJ2 — Type

OrbitPropagatorJ2{Tepoch, T} <: OrbitPropagator{Tepoch, T}

J2 orbit propagator.

Fields

j2d: Structure that stores the J2 orbit propagator data (see J2Propagator).

SatelliteToolboxPropagators.OrbitPropagatorJ2Osculating — Type

OrbitPropagatorJ2Osculating{Tepoch, T} <: OrbitPropagator{Tepoch, T}

J2 osculating orbit propagator.

Fields

j2oscd: Structure that stores the J2 osculating orbit propagator data (see J2OsculatingPropagator).

SatelliteToolboxPropagators.OrbitPropagatorJ4 — Type

OrbitPropagatorJ4{Tepoch, T} <: OrbitPropagator{Tepoch, T}

J4 orbit propagator.

Fields

j4d: Structure that stores the J4 orbit propagator data (see J4Propagator).

SatelliteToolboxPropagators.OrbitPropagatorSgp4 — Type

OrbitPropagatorSgp4{Tepoch, T} <: OrbitPropagator{Tepoch, T}

SGP4 orbit propagator.

Fields

sgp4d: Structure that stores the SGP4 orbit propagator data.

SatelliteToolboxPropagators.OrbitPropagatorTwoBody — Type

OrbitPropagatorTwoBody{Tepoch, T} <: OrbitPropagator{Tepoch, T}

Two body orbit propagator.

Fields

tbd: Structure that stores the two body orbit propagator data (see TwoBodyPropagator).

SatelliteToolboxPropagators.TwoBodyPropagator — Type

mutable struct TwoBodyPropagator{Tepoch<:Number, T<:Number}

Two body orbit propagator structure.

SatelliteToolboxPropagators.Propagators.init — Function

Propagators.init(Val(:J2), orb₀::KeplerianElements, dn_o2::Number = 0, ddn_o6::Number = 0; kwargs...) -> OrbitPropagatorJ2

Create and initialize the J2 orbit propagator structure using the mean Keplerian elements orb₀.

Note

The type used in the propagation will be the same as used to define the constants in the structure j2c.

Arguments

orb₀::KeplerianElements: Initial mean Keplerian elements [SI units].

dn_o2::Number: First time derivative of the mean motion divided by two [rad/s^2]. (Default = 0)

ddn_o6::Number: Second time derivative of the mean motion divided by six [rad/s^3]. (Default = 0)

Keywords

j2c::J2PropagatorConstants: J2 orbit propagator constants (see J2PropagatorConstants). (Default = j2c_egm2008)

SatelliteToolboxPropagators.Propagators.init — Function

Propagators.init(Val(:J4), orb₀::Orbit, dn_o4::Number = 0, ddn_o6::Number = 0; kwargs...) -> OrbitPropagatorJ4

Create and initialize the J4 orbit propagator structure using the mean Keplerian elements orb₀.

Note

The type used in the propagation will be the same as used to define the constants in the structure j4c.

Arguments

orb₀::KeplerianElements: Initial mean Keplerian elements [SI units].

dn_o4::Number: First time derivative of the mean motion divided by two [rad/s^4]. (Default = 0)

ddn_o6::Number: Second time derivative of the mean motion divided by six [rad/s^3]. (Default = 0)

Keywords

j4c::J4PropagatorConstants: J4 orbit propagator constants (see J4PropagatorConstants). (Default = j4c_egm2008)

SatelliteToolboxPropagators.Propagators.init — Function

Propagators.init(Val(:J2osc), orb₀::KeplerianElements, dn_o2::Number = 0, ddn_o6::Number = 0; kwargs...) -> OrbitPropagatorJ2Osculating

Create and initialize the J2 osculating orbit propagator structure using the mean Keplerian elements orb₀.

Note

The type used in the propagation will be the same as used to define the constants in the structure j2c.

Arguments

orb₀::KeplerianElements: Initial mean Keplerian elements [SI units].

dn_o2::Number: First time derivative of the mean motion divided by two [rad/s^2]. (Default = 0)

ddn_o6::Number: Second time derivative of the mean motion divided by six [rad/s^3]. (Default = 0)

Keywords

j2c::J2PropagatorConstants: J2 orbit propagator constants (see J2PropagatorConstants). (Default = j2c_egm2008)

SatelliteToolboxPropagators.Propagators.init! — Function

Propagators.init!(orbp::OrbitPropagatorJ2Osculating, orb₀::KeplerianElements, dn_o2::Number = 0, ddn_o6::Number = 0) -> Nothing

Initialize the J2 osculating orbit propagator structure orbp using the mean Keplerian elements orb₀.

Warning

The propagation constants j2c::J2PropagatorConstants in orbp.j2d will not be changed. Hence, they must be initialized.

Arguments

orb₀::KeplerianElements: Initial mean Keplerian elements [SI units].

dn_o2::Number: First time derivative of the mean motion divided by two [rad/s^2]. (Default = 0)

ddn_o6::Number: Second time derivative of the mean motion divided by six [rad/s^3]. (Default = 0)

SatelliteToolboxPropagators.Propagators.init! — Function

Propagators.init!(orbp::OrbitPropagatorJ4, orb₀::KeplerianElements, dn_o2::Number = 0, ddn_o6::Number = 0) -> Nothing

Initialize the J4 orbit propagator structure orbp using the mean Keplerian elements orb₀.

Warning

The propagation constants j4c::J4PropagatorConstants in j4d will not be changed. Hence, they must be initialized.

Arguments

orb₀::KeplerianElements: Initial mean Keplerian elements [SI units].

dn_o2::Number: First time derivative of the mean motion divided by two [rad/s^2]. (Default = 0)

ddn_o6::Number: Second time derivative of the mean motion divided by six [rad/s^3]. (Default = 0)

SatelliteToolboxPropagators.Propagators.init! — Function

Propagators.init!(orbp::OrbitPropagatorJ2, orb₀::KeplerianElements, dn_o2::Number = 0, ddn_o6::Number = 0) -> Nothing

Initialize the J2 orbit propagator structure orbp using the mean Keplerian elements orb₀.

Warning

The propagation constants j2c::J2PropagatorConstants in orbp.j2d will not be changed. Hence, they must be initialized.

Arguments

orb₀::KeplerianElements: Initial mean Keplerian elements [SI units].

dn_o2::Number: First time derivative of the mean motion divided by two [rad/s^2]. (Default = 0)

ddn_o6::Number: Second time derivative of the mean motion divided by six [rad/s^3]. (Default = 0)

SatelliteToolboxPropagators.Propagators.init! — Method

Propagators.init!(orbp::OrbitPropagatorSgp4, epoch::Number, n₀::Number, e₀::Number, i₀::Number, Ω₀::Number, ω₀::Number, M₀::Number, bstar::Number; kwargs...) -> Nothing
Propagators.init!(orbp::OrbitPropagatorSgp4, tle::TLE; kwargs...) -> Nothing

Initialize the SGP4 orbit propagator structure orbp using the initial orbit specified by the arguments.

Warning

The propagation constants sgp4c::Sgp4Constants in orbp.sgp4d will not be changed. Hence, they must be initialized.

Arguments

epoch::Number: Epoch of the orbital elements [Julian Day].

n₀::Number: SGP type "mean" mean motion at epoch [rad/s].

e₀::Number: "Mean" eccentricity at epoch.

i₀::Number: "Mean" inclination at epoch [rad].

Ω₀::Number: "Mean" longitude of the ascending node at epoch [rad].

ω₀::Number: "Mean" argument of perigee at epoch [rad].

M₀::Number: "Mean" mean anomaly at epoch [rad].

bstar::Number: Drag parameter (B*).

tle::TLE: Two-line elements used for the initialization.

SatelliteToolboxPropagators.Propagators.init! — Method

Propagators.init!(orbp::OrbitPropagatorTwoBody, orb₀::KeplerianElements) -> Nothing

Initialize the two-body orbit propagator structure orbp using the mean Keplerian elements orb₀.

Warning

The propagation constant μ::Number in tbd will not be changed. Hence, it must be initialized.

Arguments

orb₀::KeplerianElements: Initial mean Keplerian elements [SI units].

SatelliteToolboxPropagators.Propagators.init — Method

Propagators.init(Val(:SGP4), epoch::Number, n₀::Number, e₀::Number, i₀::Number, Ω₀::Number, ω₀::Number, M₀::Number, bstar::Number; kwargs...) -> OrbitPropagatorSgp4
Propagators.init(Val(:SGP4), tle::TLE; kwargs...) -> OrbitPropagatorSgp4

Create and initialize the SGP4 orbit propagator structure using the initial orbit specified by the arguments.

Note

The type used in the propagation will be the same as used to define the constants in the structure sgp4c.

Arguments

epoch::Number: Epoch of the orbital elements [Julian Day].

n₀::Number: SGP type "mean" mean motion at epoch [rad/s].

e₀::Number: "Mean" eccentricity at epoch.

i₀::Number: "Mean" inclination at epoch [rad].

Ω₀::Number: "Mean" longitude of the ascending node at epoch [rad].

ω₀::Number: "Mean" argument of perigee at epoch [rad].

M₀::Number: "Mean" mean anomaly at epoch [rad].

bstar::Number: Drag parameter (B*).

tle::TLE: Two-line elements used for the initialization.

Keywords

sgp4c::Sgp4Constants: SGP4 orbit propagator constants (see Sgp4Constants). (Default = sgp4c_wgs84)

SatelliteToolboxPropagators.Propagators.init — Method

Propagators.init(Val(:TwoBody), orb₀::KeplerianElements; kwargs...) -> OrbitPropagatorTwoBody

Create and initialize the two-body orbit propagator structure using the mean Keplerian elements orb₀.

Note

The type used in the propagation will be the same as used to define the gravitational constant μ.

Arguments

orb₀::KeplerianElements: Initial mean Keplerian elements [SI units].

Keywords

μ::T: Standard gravitational parameter of the central body [m³/s²]. (Default = tbc_m0)

SatelliteToolboxPropagators.Propagators.propagate — Function

Propagators.propagate(Val(:J2osc), Δt::Number, orb₀::KeplerianElements, dn_o2::Number = 0, ddn_o6::Number = 0; kwargs...)

Initialize the J2 osculating propagator structure using the input elements orb₀ and propagate the orbit until the time Δt [s].

Note

The type used in the propagation will be the same as used to define the constants in the structure j2c.

Arguments

orb₀::KeplerianElements: Initial mean Keplerian elements [SI units].

dn_o2::Number: First time derivative of the mean motion divided by two [rad/s^2]. (Default = 0)

ddn_o6::Number: Second time derivative of the mean motion divided by six [rad/s^3]. (Default = 0)

Keywords

j2c::J2PropagatorConstants{T}: J2 orbit propagator constants (see J2PropagatorConstants). (Default = j2c_egm2008)

Returns

SVector{3, T}: Position vector [m] represented in the inertial frame at propagation instant.

SVector{3, T}: Velocity vector [m / s] represented in the inertial frame at propagation instant.

OrbitPropagatorJ2Osculating: Structure with the initialized parameters.

SatelliteToolboxPropagators.Propagators.propagate — Function

Propagators.propagate(Val(:J4), Δt::Number, orb₀::KeplerianElements, dn_o2::Number = 0, ddn_o6::Number = 0; kwargs...)

Initialize the J4 propagator structure using the input elements orb₀ and propagate the orbit until the time Δt [s].

Note

The type used in the propagation will be the same as used to define the constants in the structure j4c.

Arguments

orb₀::KeplerianElements: Initial mean Keplerian elements [SI units].

dn_o2::Number: First time derivative of the mean motion divided by two [rad/s^4]. (Default = 0)

ddn_o6::Number: Second time derivative of the mean motion divided by six [rad/s^3]. (Default = 0)

Keywords

j4c::J4PropagatorConstants{T}: J4 orbit propagator constants (see J4PropagatorConstants). (Default = j4c_egm2008)

Returns

SVector{3, T}: Position vector [m] represented in the inertial frame at propagation instant.

SVector{3, T}: Velocity vector [m / s] represented in the inertial frame at propagation instant.

OrbitPropagatorJ4: Structure with the initialized parameters.

SatelliteToolboxPropagators.Propagators.propagate — Function

Propagators.propagate(Val(:J2), Δt::Number, orb₀::KeplerianElements, dn_o2::Number = 0, ddn_o6::Number = 0; kwargs...)

Initialize the J2 propagator structure using the input elements orb₀ and propagate the orbit until the time Δt [s].

Note

The type used in the propagation will be the same as used to define the constants in the structure j2c.

Arguments

orb₀::KeplerianElements: Initial mean Keplerian elements [SI units].

dn_o2::Number: First time derivative of the mean motion divided by two [rad/s^2]. (Default = 0)

ddn_o6::Number: Second time derivative of the mean motion divided by six [rad/s^3]. (Default = 0)

Keywords

j2c::J2PropagatorConstants{T}: J2 orbit propagator constants (see J2PropagatorConstants). (Default = j2c_egm2008)

Returns

SVector{3, T}: Position vector [m] represented in the inertial frame at propagation instant.

SVector{3, T}: Velocity vector [m / s] represented in the inertial frame at propagation instant.

OrbitPropagatorJ2: Structure with the initialized parameters.

SatelliteToolboxPropagators.Propagators.propagate — Method

Propagators.propagate(Val(:SGP4), Δt::Number, epoch::Number, n₀::Number, e₀::Number, i₀::Number, Ω₀::Number, ω₀::Number, M₀::Number, bstar::Number; kwargs...)
Propagators.propagate(Val(:SGP4), Δt::Number, tle::TLE; kwargs...)

Initialize the SGP4 propagator structure using the input arguments and propagate the orbit until the time Δt [s].

Note

The type used in the propagation will be the same as used to define the constants in the structure sgp4c.

Arguments

epoch::Number: Epoch of the orbital elements [Julian Day].

n₀::Number: SGP type "mean" mean motion at epoch [rad/s].

e₀::Number: "Mean" eccentricity at epoch.

i₀::Number: "Mean" inclination at epoch [rad].

Ω₀::Number: "Mean" longitude of the ascending node at epoch [rad].

ω₀::Number: "Mean" argument of perigee at epoch [rad].

M₀::Number: "Mean" mean anomaly at epoch [rad].

bstar::Number: Drag parameter (B*).

tle::TLE: Two-line elements used for the initialization.

Keywords

sgp4c::Sgp4Constants{T}: SGP4 orbit propagator constants (see Sgp4Constants). (Default = sgp4c_wgs84)

Returns

SVector{3, T}: Position vector [m] represented in the inertial frame at propagation instant.

SVector{3, T}: Velocity vector [m / s] represented in the inertial frame at propagation instant.

OrbitPropagatorSgp4: Structure with the initialized parameters.

SatelliteToolboxPropagators.Propagators.propagate — Method

Propagators.propagate(Val(:TwoBody), Δt::Number, orb₀::KeplerianElements; kwargs...)

Initialize the two-body propagator structure using the input elements orb₀ and propagate the orbit until the time Δt [s].

Note

The type used in the propagation will be the same as used to define the gravitational constant μ.

Arguments

orb₀::KeplerianElements: Initial mean Keplerian elements [SI units].

Keywords

μ::T: Standard gravitational parameter of the central body [m³/s²]. (Default = tbc_m0)

Returns

SVector{3, T}: Position vector [m] represented in the inertial frame at propagation instant.

SVector{3, T}: Velocity vector [m / s] represented in the inertial frame at propagation instant.

TwoBodyPropagator: Structure with the initialized parameters.

SatelliteToolboxPropagators.j2! — Method

j2!(j2d::J2Propagator{Tepoch, T}, t::Number) where {Tepoch, T} -> SVector{3, T}, SVector{3, T}

Propagate the orbit defined in j2d (see J2Propagator) to t [s] after the epoch of the input mean elements in j2d.

Note

The internal values in j2d will be modified.

Returns

SVector{3, T}: Position vector [m] represented in the inertial frame at propagation instant.

SVector{3, T}: Velocity vector [m / s] represented in the inertial frame at propagation instant.

Remarks

The inertial frame in which the output is represented depends on which frame it was used to generate the orbit parameters. Notice that the perturbation theory requires an inertial frame with true equator.

SatelliteToolboxPropagators.j2 — Method

j2(Δt::Number, orb₀::KeplerianElements, dn_o2::Number = 0, ddn_o6::Number = 0; kwargs...)

Initialize the J2 propagator structure using the input elements orb₀ and propagate the orbit until the time Δt [s].

Note

The type used in the propagation will be the same as used to define the constants in the structure j2c.

Arguments

orb₀::KeplerianElements: Initial mean Keplerian elements [SI units].

dn_o2::Number: First time derivative of the mean motion divided by two [rad/s^2]. (Default = 0)

ddn_o6::Number: Second time derivative of the mean motion divided by six [rad/s^3]. (Default = 0)

Keywords

j2c::J2PropagatorConstants{T}: J2 orbit propagator constants (see J2PropagatorConstants). (Default = j2c_egm2008)

Returns

SVector{3, T}: Position vector [m] represented in the inertial frame at propagation instant.

SVector{3, T}: Velocity vector [m / s] represented in the inertial frame at propagation instant.

J2Propagator: Structure with the initialized parameters.

Remarks

The inertial frame in which the output is represented depends on which frame it was used to generate the orbit parameters. Notice that the perturbation theory requires an inertial frame with true equator.

SatelliteToolboxPropagators.j2_init! — Method

j2_init!(j2d::J2Propagator, orb₀::KeplerianElements, dn_o2::Number = 0, ddn_o6::Number = 0) -> Nothing

Initialize the J2 orbit propagator structure j2d using the mean Keplerian elements orb₀.

Warning

The propagation constants j2c::J2PropagatorConstants in j2d will not be changed. Hence, they must be initialized.

Arguments

orb₀::KeplerianElements: Initial mean Keplerian elements [SI units].

dn_o2::Number: First time derivative of the mean motion divided by two [rad/s^2]. (Default = 0)

ddn_o6::Number: Second time derivative of the mean motion divided by six [rad/s^3]. (Default = 0)

SatelliteToolboxPropagators.j2_init — Method

j2_init(orb₀::KeplerianElements, dn_o2::Number = 0, ddn_o6::Number = 0; kwargs...) where T<:Number -> J2Propagator

Create and initialize the J2 orbit propagator structure using the mean Keplerian elements orb₀.

Note

The type used in the propagation will be the same as used to define the constants in the structure j2c.

Arguments

orb₀::KeplerianElements: Initial mean Keplerian elements [SI units].

dn_o2::Number: First time derivative of the mean motion divided by two [rad/s^2]. (Default = 0)

ddn_o6::Number: Second time derivative of the mean motion divided by six [rad/s^3]. (Default = 0)

Keywords

j2c::J2PropagatorConstants: J2 orbit propagator constants (see J2PropagatorConstants). (Default = j2c_egm2008)

SatelliteToolboxPropagators.j2osc! — Method

j2osc!(j2oscd::J2OsculatingPropagator{Tepoch, T}, t::Number) where {Tepoch, T} -> SVector{3, T}, SVector{3, T}

Propagate the orbit defined in j2oscd (see J2OsculatingPropagator) to t [s] after the epoch of the input mean elements in j2d.

Note

The internal values in j2oscd will be modified.

Returns

SVector{3, T}: Position vector [m] represented in the inertial frame at propagation instant.

SVector{3, T}: Velocity vector [m / s] represented in the inertial frame at propagation instant.

Remarks

The inertial frame in which the output is represented depends on which frame it was used to generate the orbit parameters. Notice that the perturbation theory requires an inertial frame with true equator.

SatelliteToolboxPropagators.j2osc — Method

j2osc(Δt::Number, orb₀::KeplerianElements, dn_o2::Number = 0, ddn_o6::Number = 0; kwargs...)

Initialize the J2 osculating propagator structure using the input elements orb₀ and propagate the orbit until the time Δt [s].

Note

The type used in the propagation will be the same as used to define the constants in the structure j2c.

Arguments

orb₀::KeplerianElements: Initial mean Keplerian elements [SI units].

dn_o2::Number: First time derivative of the mean motion divided by two [rad/s^2]. (Default = 0)

ddn_o6::Number: Second time derivative of the mean motion divided by six [rad/s^3]. (Default = 0)

Keywords

j2c::J2PropagatorConstants{T}: J2 orbit propagator constants (see J2PropagatorConstants). (Default = j2c_egm2008)

Returns

SVector{3, T}: Position vector [m] represented in the inertial frame at propagation instant.

SVector{3, T}: Velocity vector [m / s] represented in the inertial frame at propagation instant.

J2OsculatingPropagator: Structure with the initialized parameters.

Remarks

The inertial frame in which the output is represented depends on which frame it was used to generate the orbit parameters. Notice that the perturbation theory requires an inertial frame with true equator.

SatelliteToolboxPropagators.j2osc_init! — Method

j2osc_init!(j2oscd::J2OsculatingPropagator, orb₀::KeplerianElements, dn_o2::Number = 0, ddn_o6::Number = 0) -> Nothing

Initialize the J2 osculating orbit propagator structure j2oscd using the mean Keplerian elements orb₀.

Warning

The propagation constants j2c::J2PropagatorConstants in j2oscd.j2d will not be changed. Hence, they must be initialized.

Arguments

orb₀::KeplerianElements: Initial mean Keplerian elements [SI units].

dn_o2::Number: First time derivative of the mean motion divided by two [rad/s^2]. (Default = 0)

ddn_o6::Number: Second time derivative of the mean motion divided by six [rad/s^3]. (Default = 0)

SatelliteToolboxPropagators.j2osc_init — Method

j2osc_init(orb₀::KeplerianElements, dn_o2::Number = 0, ddn_o6::Number = 0; kwargs...) where T<:Number -> J2OsculatingPropagator

Create and initialize the J2 osculating orbit propagator structure using the mean Keplerian elements orb₀.

Note

The type used in the propagation will be the same as used to define the constants in the structure j2c.

Arguments

orb₀::KeplerianElements: Initial mean Keplerian elements [SI units].

dn_o2::Number: First time derivative of the mean motion divided by two [rad/s^2]. (Default = 0)

ddn_o6::Number: Second time derivative of the mean motion divided by six [rad/s^3]. (Default = 0)

Keywords

j2c::J2PropagatorConstants: J2 orbit propagator constants (see J2PropagatorConstants). (Default = j2c_egm2008)

SatelliteToolboxPropagators.j4! — Method

j4!(j4d::J4Propagator{Tepoch, T}, t::Number) where {Tepoch, T} -> SVector{3, T}, SVector{3, T}

Propagate the orbit defined in j4d (see J4Propagator) to t [s] after the epoch of the input mean elements in j4d.

Note

The internal values in j4d will be modified.

Returns

SVector{3, T}: Position vector [m] represented in the inertial frame at propagation instant.

SVector{3, T}: Velocity vector [m / s] represented in the inertial frame at propagation instant.

Remarks

The inertial frame in which the output is represented depends on which frame it was used to generate the orbit parameters. Notice that the perturbation theory requires an inertial frame with true equator.

SatelliteToolboxPropagators.j4 — Method

j4(Δt::Number, orb₀::KeplerianElements, dn_o2::Number = 0, ddn_o6::Number = 0; kwargs...)

Initialize the J4 propagator structure using the input elements orb₀ and propagate the orbit until the time Δt [s].

Note

The type used in the propagation will be the same as used to define the constants in the structure j4c.

Arguments

orb₀::KeplerianElements: Initial mean Keplerian elements [SI units].

dn_o2::Number: First time derivative of the mean motion divided by two [rad/s^4]. (Default = 0)

ddn_o6::Number: Second time derivative of the mean motion divided by six [rad/s^3]. (Default = 0)

Keywords

j4c::J4PropagatorConstants{T}: J4 orbit propagator constants (see J4PropagatorConstants). (Default = j4c_egm2008)

Returns

SVector{3, T}: Position vector [m] represented in the inertial frame at propagation instant.

SVector{3, T}: Velocity vector [m / s] represented in the inertial frame at propagation instant.

J4Propagator: Structure with the initialized parameters.

Remarks

The inertial frame in which the output is represented depends on which frame it was used to generate the orbit parameters. Notice that the perturbation theory requires an inertial frame with true equator.

SatelliteToolboxPropagators.j4_init! — Method

j4_init!(j4d::J4Propagator, orb₀::KeplerianElements, dn_o4::Number = 0, ddn_o6::Number = 0) -> Nothing

Initialize the J4 orbit propagator structure j4d using the mean Keplerian elements orb₀.

Warning

The propagation constants j4c::J4PropagatorConstants in j4d will not be changed. Hence, they must be initialized.

Arguments

orb₀::KeplerianElements: Initial mean Keplerian elements [SI units].

dn_o2::Number: First time derivative of the mean motion divided by two [rad/s^2]. (Default = 0)

ddn_o6::Number: Second time derivative of the mean motion divided by six [rad/s^3]. (Default = 0)

SatelliteToolboxPropagators.j4_init — Method

j4_init(orb₀::KeplerianElements, dn_o2::Number = 0, ddn_o6::Number = 0; kwargs...) where T<:Number -> J4Propagator

Create and initialize the J4 orbit propagator structure using the mean Keplerian elements orb₀.

Note

The type used in the propagation will be the same as used to define the constants in the structure j4c.

Arguments

orb₀::KeplerianElements: Initial mean Keplerian elements [SI units].

dn_o2::Number: First time derivative of the mean motion divided by two [rad/s^2]. (Default = 0)

ddn_o6::Number: Second time derivative of the mean motion divided by six [rad/s^3]. (Default = 0)

Keywords

j4c::J4PropagatorConstants: J4 orbit propagator constants (see J4PropagatorConstants). (Default = j4c_egm2008)

SatelliteToolboxPropagators.twobody! — Method

twobody!(tbd::TwoBodyPropagator{Tepoch, T}, t::Number) where {Tepoch, T}

Propagate the orbit defined in tbd (see TwoBodyPropagator) to t [s] after the epoch of the input mean elements in tbd.

Note

The internal values in tbd will be modified.

Returns

SVector{3, T}: Position vector [m] represented in the inertial frame at propagation instant.

SVector{3, T}: Velocity vector [m / s] represented in the inertial frame at propagation instant.

Remarks

The inertial frame in which the output is represented depends on which frame it was used to generate the orbit parameters. Notice that the perturbation theory requires an inertial frame with true equator.

SatelliteToolboxPropagators.twobody — Method

twobody(Δt::Number, orb₀::KeplerianElements; kwargs...)

Initialize the two-body propagator structure using the input elements orb₀ and propagate the orbit until the time Δt [s].

Note

The type used in the propagation will be the same as used to define the gravitational constant μ.

Arguments

orb₀::KeplerianElements: Initial mean Keplerian elements [SI units].

Keywords

μ::T: Standard gravitational parameter of the central body [m³/s²]. (Default = tbc_m0)

Returns

SVector{3, T}: Position vector [m] represented in the inertial frame at propagation instant.

SVector{3, T}: Velocity vector [m / s] represented in the inertial frame at propagation instant.

TwoBodyPropagator: Structure with the initialized parameters.

Remarks

The inertial frame in which the output is represented depends on which frame it was used to generate the orbit parameters. Notice that the perturbation theory requires an inertial frame with true equator.

SatelliteToolboxPropagators.twobody_init! — Method

twobody_init!(tbd::TwoBodyPropagator, orb₀::KeplerianElements; kwargs...) -> Nothing

Initialize the two-body propagator structure tbd using the mean Keplerian elements orb₀.

Warning

The propagation constant μ::Number in tbd will not be changed. Hence, it must be initialized.

Arguments

orb₀::KeplerianElements: Initial mean Keplerian elements [SI units].

SatelliteToolboxPropagators.twobody_init — Method

twobody_init(orb₀::KeplerianElements; kwargs...) -> TwoBodyPropagator

Create and initialize the two-body propagator structure using the mean Keplerian elements orb₀.

Note

The type used in the propagation will be the same as used to define the gravitational constant μ.

Arguments

orb₀::KeplerianElements: Initial mean Keplerian elements [SI units].

Keywords

μ::T: Standard gravitational parameter of the central body [m³/s²]. (Default = tbc_m0)